Rename refresh command to pull. Create pull sub command stubs.
