+++ src/Command/Pull/PullCommandBase.php
@@ -1,135 +1,29 @@
 <?php
 
-namespace Acquia\Cli\Command;
+namespace Acquia\Cli\Command\Pull;
 
+use Acquia\Cli\Command\CommandBase;
 use Acquia\Cli\Exception\AcquiaCliException;
-use Acquia\Cli\Output\Checklist;
 use Acquia\DrupalEnvironmentDetector\AcquiaDrupalEnvironmentDetector;
 use AcquiaCloudApi\Connector\Client;
 use AcquiaCloudApi\Endpoints\Environments;
-use AcquiaCloudApi\Exception\ApiErrorException;
 use AcquiaCloudApi\Response\EnvironmentResponse;
-use stdClass;
 use Symfony\Component\Console\Input\ArrayInput;
-use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Question\ChoiceQuestion;
 use Symfony\Component\Console\Question\ConfirmationQuestion;
-use Symfony\Component\Process\Process;
 use Webmozart\PathUtil\Path;
 
 /**
- * Class RefreshCommand.
+ * Class PullCommandBase.
  */
-class RefreshCommand extends CommandBase {
-
-  protected static $defaultName = 'refresh';
+abstract class PullCommandBase extends CommandBase {
 
   /**
    * @var string
    */
   protected $dir;
-
-  /**
-   * {inheritdoc}.
-   */
-  protected function configure() {
-    $this->setName('refresh')
-      ->setDescription('Copy code, database, and files from a Cloud Platform environment')
-      ->addArgument('dir', InputArgument::OPTIONAL, 'The directory containing the Drupal project to be refreshed')
-      ->addOption('cloud-env-uuid', 'from', InputOption::VALUE_REQUIRED, 'The UUID of the associated Cloud Platform source environment')
-      ->addOption('no-code', NULL, InputOption::VALUE_NONE, 'Do not refresh code from remote repository')
-      ->addOption('no-files', NULL, InputOption::VALUE_NONE, 'Do not refresh files')
-      ->addOption('no-databases', NULL, InputOption::VALUE_NONE, 'Do not refresh databases')
-      ->addOption(
-            'no-scripts',
-            NULL,
-            InputOption::VALUE_NONE,
-            'Do not run any additional scripts after code and database are copied. E.g., composer install , drush cache-rebuild, etc.'
-        )
-      ->addOption('scripts', NULL, InputOption::VALUE_NONE, 'Only execute additional scripts');
-  }
-
-  /**
-   * @param \Symfony\Component\Console\Input\InputInterface $input
-   * @param \Symfony\Component\Console\Output\OutputInterface $output
-   *
-   * @return int 0 if everything went fine, or an exit code
-   * @throws \Exception
-   */
-  protected function execute(InputInterface $input, OutputInterface $output) {
-    $this->determineDir($input);
-    if ($this->dir !== '/home/ide/project' && AcquiaDrupalEnvironmentDetector::isAhIdeEnv()) {
-      throw new AcquiaCliException('Please run this command from the {dir} directory', ['dir' => '/home/ide/project']);
-    }
-
-    $clone = $this->determineCloneProject($output);
-    $acquia_cloud_client = $this->cloudApiClientService->getClient();
-    $chosen_environment = $this->determineEnvironment($input, $output, $acquia_cloud_client);
-    $checklist = new Checklist($output);
-    $output_callback = static function ($type, $buffer) use ($checklist, $output) {
-      if (!$output->isVerbose()) {
-        $checklist->updateProgressBar($buffer);
-      }
-      $output->writeln($buffer, OutputInterface::VERBOSITY_VERY_VERBOSE);
-    };
-
-    if (!$input->getOption('no-code')) {
-      if ($clone) {
-        $checklist->addItem('Cloning git repository from the Cloud Platform');
-        $this->cloneFromCloud($chosen_environment, $output_callback);
-        $checklist->completePreviousItem();
-      }
-      else {
-        $checklist->addItem('Pulling code from the Cloud Platform');
-        $this->pullCodeFromCloud($chosen_environment, $output_callback);
-        $checklist->completePreviousItem();
-      }
-    }
-
-    // Copy databases.
-    if (!$input->getOption('no-databases')) {
-      $database = $this->determineSourceDatabase($acquia_cloud_client, $chosen_environment);
-      $checklist->addItem('Importing Drupal database copy from the Cloud Platform');
-      $this->importRemoteDatabase($chosen_environment, $database, $output_callback);
-      $checklist->completePreviousItem();
-    }
-
-    // Copy files.
-    if (!$input->getOption('no-files')) {
-      $checklist->addItem('Copying Drupal\'s public files from the Cloud Platform');
-      $this->rsyncFilesFromCloud($chosen_environment, $output_callback);
-      $checklist->completePreviousItem();
-    }
-
-    if (!$input->getOption('no-scripts')) {
-      if (file_exists($this->dir . '/composer.json') && $this->localMachineHelper
-        ->commandExists('composer')) {
-        $checklist->addItem('Installing Composer dependencies');
-        $this->composerInstall($output_callback);
-        $checklist->completePreviousItem();
-      }
-
-      if ($this->drushHasActiveDatabaseConnection($output_callback)) {
-        // Drush rebuild caches.
-        $checklist->addItem('Clearing Drupal caches via Drush');
-        $this->drushRebuildCaches($output_callback);
-        $checklist->completePreviousItem();
-
-        // Drush sanitize.
-        $checklist->addItem('Sanitizing database via Drush');
-        $this->drushSqlSanitize($output_callback);
-        $checklist->completePreviousItem();
-      }
-    }
-
-    // Match IDE PHP version to source environment PHP version.
-    $this->matchIdePhpVersion($output, $chosen_environment);
-
-    return 0;
-  }
 
   /**
    * @return bool
@@ -402,9 +296,9 @@ class RefreshCommand extends CommandBase {
     // Re-key the array since we removed production.
     $application_environments = array_values($application_environments);
     $question = new ChoiceQuestion(
-          '<question>Choose a Cloud Platform environment to copy from</question>:',
-          $choices
-      );
+      '<question>Choose a Cloud Platform environment to copy from</question>:',
+      $choices
+    );
     $helper = $this->getHelper('question');
     $chosen_environment_label = $helper->ask($this->input, $this->output, $question);
     $chosen_environment_index = array_search($chosen_environment_label, $choices, TRUE);
@@ -447,10 +341,10 @@ class RefreshCommand extends CommandBase {
     }
 
     $question = new ChoiceQuestion(
-          '<question>Choose a database to copy</question>:',
-          $choices,
-          $default_database_index
-      );
+      '<question>Choose a database to copy</question>:',
+      $choices,
+      $default_database_index
+    );
     $helper = $this->getHelper('question');
     $chosen_database_label = $helper->ask($this->input, $this->output, $question);
     $chosen_database_index = array_search($chosen_database_label, $choices, TRUE);
@@ -567,11 +461,11 @@ class RefreshCommand extends CommandBase {
    * @return object
    * @throws \Exception
    */
-  protected function determineSourceDatabase(Client $acquia_cloud_client, $chosen_environment): stdClass {
+  protected function determineSourceDatabase(Client $acquia_cloud_client, $chosen_environment): \stdClass {
     $databases = $acquia_cloud_client->request(
-          'get',
-          '/environments/' . $chosen_environment->uuid . '/databases'
-      );
+      'get',
+      '/environments/' . $chosen_environment->uuid . '/databases'
+    );
     if (count($databases) > 1) {
       $database = $this->promptChooseDatabase($chosen_environment, $databases);
     }
